Email Parse New Hire Account Creation with GAS
This script was created as a temporary solution to avoid manual user account creation using Google apps script and Google Admin SDK. It was determined that Google Apps Script was the best interim solution due to access limitation during the HRIS transition. The New HRIS would send automated emails when a user completed all onboarding actions and this script would parse those emails to find new hire accounts that needed to be created each day. This was the final version of the script.